+++ kikuchipy/data/_registry.py
@@ -21,8 +21,14 @@ registry = {
     "data/kikuchipy/patterns.h5": "7a99ce88174c725f5407b6fcc1eab0c4255694ca8e6029fde7f372f3ab40897f",
     "data/emsoft_ebsd_master_pattern/ni_mc_mp_20kv_uint8_gzip_opts9.h5": "8a7c1fb471d9ce750f0332a154e87cf41eed7529be508548e0c0f51ec6f92bc2",
     "data/nickel_ebsd_large/patterns.h5": "3ea6e729c3adfdea9dce461806f011c24bf70b011dcf4d90a23a6aa29f15872c",
+    "data/silicon_ebsd_moving_screen/si_in.h5": "de57b8ef7213420af694c191c939f44818979b3db9873e74cc85cabde824b7eb",
+    "data/silicon_ebsd_moving_screen/si_out5mm.h5": "5d5a53a19f87316c2f20105365660d197302ad8c5e7ec4f1c75d024e787b0a6d",
+    "data/silicon_ebsd_moving_screen/si_out10mm.h5": "540b9dd485b9177741f2175d6225040419b8b585a42da43011fbe86c7a5d26dd",
 }
 registry_urls = {
-    "data/nickel_ebsd_large/patterns.h5": "https://github.com/pyxem/kikuchipy-data/raw/master/nickel_ebsd_large/patterns.h5"
+    "data/nickel_ebsd_large/patterns.h5": "https://github.com/pyxem/kikuchipy-data/raw/master/nickel_ebsd_large/patterns.h5",
+    "data/silicon_ebsd_moving_screen/si_in.h5": "https://github.com/pyxem/kikuchipy-data/raw/master/silicon_ebsd_moving_screen/si_in.h5",
+    "data/silicon_ebsd_moving_screen/si_out5mm.h5": "https://github.com/pyxem/kikuchipy-data/raw/master/silicon_ebsd_moving_screen/si_out5mm.h5",
+    "data/silicon_ebsd_moving_screen/si_out10mm.h5": "https://github.com/pyxem/kikuchipy-data/raw/master/silicon_ebsd_moving_screen/si_out10mm.h5",
 }
 # fmt: on